% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_notifications_for_budget}
\alias{budgets_describe_notifications_for_budget}
\title{Lists the notifications that are associated with a budget}
\usage{
budgets_describe_notifications_for_budget(AccountId, BudgetName,
  MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget whose notifications
you want descriptions of.}

\item{BudgetName}{[required] The name of the budget whose notifications you want descriptions of.}

\item{MaxResults}{An optional integer that represents how many entries a paginated
response contains. The maximum is 100.}

\item{NextToken}{The pagination token that you include in your request to indicate the
next set of results that you want to retrieve.}
}
\description{
Lists the notifications that are associated with a budget.
}
\section{Request syntax}{
\preformatted{svc$describe_notifications_for_budget(
  AccountId = "string",
  BudgetName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
